added show content functionality and error handling to list notes

+++ notes.py
@@ -4,11 +4,15 @@ notes_dir = config["notes_dir"]
 
 
 def create_new_note(name: str):
+    if os.path.exists(f"{notes_dir}/{name}.txt"):
+        raise ValueError("Note already exists.")
     with open(f"{notes_dir}/{name}.txt", 'w') as f:
-        f.close()
+        f.write("")
 
 
 def add_content_to_note(name: str, content: str):
+    if not os.path.exists(f"{notes_dir}/{name}.txt"):
+        raise ValueError("Note does not exist.")
     with open(f"{notes_dir}/{name}.txt", 'a') as f:
         f.write(content)
 
@@ -19,3 +23,11 @@ def list_all_notes() -> list:
         if file.endswith('.txt'):
             files.append(file)
     return files
+
+
+def get_notes_content(name: str) -> str:
+    if not os.path.exists(f"{notes_dir}/{name}.txt"):
+        raise ValueError("Note does not exist.")
+    with open(f"{notes_dir}/{name}.txt", 'r') as file:
+        notes_data = file.read()
+    return notes_data
